refactor[#11] : database 타입 shared/models 로 이동
