+++ blog/src/shared/model/database.types.ts
